backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
pymupdf
Pillow
numpy
opencv-python
groq
python-dotenv
torch
torchvision
httpx
beautifulsoup4
